Refactor schemas
