+++ src/quacc/recipes/dftb/_base.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 from ase.calculators.dftb import Dftb
 
 from quacc.runners.ase import Runner
-from quacc.schemas.ase import summarize_run
+from quacc.schemas.ase import Summarize
 from quacc.utils.dicts import recursive_dict_merge
 
 if TYPE_CHECKING:
@@ -50,7 +50,7 @@ def run_and_summarize(
     Returns
     -------
     RunSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_run][]
+        Dictionary of results, specified in [quacc.schemas.ase.Summarize.run][]
     """
     calc_flags = recursive_dict_merge(calc_defaults, calc_swaps)
 
@@ -59,4 +59,6 @@ def run_and_summarize(
         geom_file=GEOM_FILE
     )
 
-    return summarize_run(final_atoms, atoms, additional_fields=additional_fields)
+    return Summarize(additional_fields=additional_fields).run(
+        final_atoms, atoms
+    )
