+++ src/quacc/recipes/dftb/core.py
@@ -45,7 +45,7 @@ def static_job(
     Returns
     -------
     RunSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_run][].
+        Dictionary of results, specified in [quacc.schemas.ase.Summarize.run][].
         See the return type-hint for the data structure.
     """
     calc_defaults = {
@@ -99,7 +99,7 @@ def relax_job(
     Returns
     -------
     RunSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_run][].
+        Dictionary of results, specified in [quacc.schemas.ase.Summarize.run][].
         See the return type-hint for the data structure.
     """
     calc_defaults = {
