+++ src/quacc/recipes/emt/defects.py
@@ -95,7 +95,7 @@ def bulk_to_defects_flow(
     Returns
     -------
     list[RunSchema | OptSchema]
-        List of dictionary of results from [quacc.schemas.ase.summarize_run][]
+        List of dictionary of results from [quacc.schemas.ase.Summarize.run][]
         or [quacc.schemas.ase.summarize_opt_run][].
         See the return type-hint for the data structure.
     """
