+++ src/quacc/recipes/lj/core.py
@@ -13,7 +13,8 @@ from ase.calculators.lj import LennardJones
 from quacc import job
 from quacc.runners.ase import Runner
 from quacc.runners.thermo import ThermoRunner
-from quacc.schemas.ase import summarize_opt_run, summarize_run, summarize_vib_and_thermo
+from quacc.schemas.ase import Summarize
+from quacc.schemas.thermo import summarize_vib_and_thermo
 
 if TYPE_CHECKING:
     from ase.atoms import Atoms
@@ -52,13 +53,13 @@ def static_job(
     Returns
     -------
     RunSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_run][].
+        Dictionary of results, specified in [quacc.schemas.ase.Summarize.run][].
         See the type-hint for the data structure.
     """
     calc = LennardJones(**calc_kwargs)
     final_atoms = Runner(atoms, calc, copy_files=copy_files).run_calc()
 
-    return summarize_run(final_atoms, atoms, additional_fields={"name": "LJ Static"})
+    return Summarize(additional_fields={"name": "LJ Static"}).run(final_atoms, atoms)
 
 
 @job
@@ -88,7 +89,7 @@ def relax_job(
     Returns
     -------
     OptSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_run][].
+        Dictionary of results, specified in [quacc.schemas.ase.Summarize.run][].
         See the type-hint for the data structure.
     """
     opt_params = opt_params or {}
@@ -96,7 +97,7 @@ def relax_job(
     calc = LennardJones(**calc_kwargs)
     dyn = Runner(atoms, calc, copy_files=copy_files).run_opt(**opt_params)
 
-    return summarize_opt_run(dyn, additional_fields={"name": "LJ Relax"})
+    return Summarize(additional_fields={"name": "LJ Relax"}).opt(dyn)
 
 
 @job
@@ -134,7 +135,7 @@ def freq_job(
     Returns
     -------
     VibThermoSchema
-        Dictionary of results, specified in [quacc.schemas.ase.summarize_vib_and_thermo][].
+        Dictionary of results, specified in [quacc.schemas.thermo.summarize_vib_and_thermo][].
         See the type-hint for the data structure.
     """
     vib_kwargs = vib_kwargs or {}
