+++ src/quacc/recipes/qchem/_base.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from quacc.calculators.qchem import QChem
 from quacc.runners.ase import Runner
-from quacc.schemas.ase import summarize_opt_run, summarize_run
+from quacc.schemas.ase import Summarize
 from quacc.utils.dicts import recursive_dict_merge
 
 if TYPE_CHECKING:
@@ -52,7 +52,7 @@ def run_and_summarize(
     Returns
     -------
     RunSchema
-        Dictionary of results from [quacc.schemas.ase.summarize_run][]
+        Dictionary of results from [quacc.schemas.ase.Summarize.run][]
     """
     calc_flags = recursive_dict_merge(calc_defaults, calc_swaps)
     calc = QChem(
@@ -60,12 +60,10 @@ def run_and_summarize(
     )
     final_atoms = Runner(atoms, calc, copy_files=copy_files).run_calc()
 
-    return summarize_run(
-        final_atoms,
-        atoms,
+    return Summarize(
         charge_and_multiplicity=(charge, spin_multiplicity),
         additional_fields=additional_fields,
-    )
+    ).run(final_atoms, atoms)
 
 
 def run_and_summarize_opt(
@@ -116,8 +114,7 @@ def run_and_summarize_opt(
     )
     dyn = Runner(atoms, calc, copy_files=copy_files).run_opt(**opt_flags)
 
-    return summarize_opt_run(
-        dyn,
+    return Summarize(
         charge_and_multiplicity=(charge, spin_multiplicity),
         additional_fields=additional_fields,
-    )
+    ).opt(dyn)
